Extract default OMA interface into separate module
Prepare to support Orthology MAtric and OrthoDB
